chore move form and request namespace
