+++ src/Siklid/Application/Auth/RegisterByEmail.php
@@ -6,9 +6,9 @@ namespace App\Siklid\Application\Auth;
 
 use App\Foundation\Action\AbstractAction;
 use App\Foundation\Security\Token\TokenManagerInterface;
-use App\Siklid\Application\Auth\Form\UserType;
-use App\Siklid\Application\Auth\Request\RegisterByEmailRequest as Request;
 use App\Siklid\Document\User;
+use App\Siklid\Form\UserType;
+use App\Siklid\Request\Auth\RegisterByEmailRequest as Request;
 use Doctrine\ODM\MongoDB\DocumentManager;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface as Hash;
