+++ src/Siklid/Request/Auth/RegisterByEmailRequest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Siklid\Application\Auth\Request;
+namespace App\Siklid\Request\Auth;
 
 use App\Foundation\Http\Request;
 
@@ -10,6 +10,6 @@ final class RegisterByEmailRequest extends Request
 {
     public function formInput(): array
     {
-        return (array) ($this->all()['user'] ?? []);
+        return (array)($this->all()['user'] ?? []);
     }
 }
